+++ pkg/api/response/lease.go
@@ -10,6 +10,7 @@ func CreateLeaseResponse(redboxLease *db.RedboxLease) *LeaseResponse {
 	return &LeaseResponse{
 		AccountID:                redboxLease.AccountID,
 		PrincipalID:              redboxLease.PrincipalID,
+		ID:                       redboxLease.ID,
 		LeaseStatus:              redboxLease.LeaseStatus,
 		CreatedOn:                redboxLease.CreatedOn,
 		LastModifiedOn:           redboxLease.LastModifiedOn,
@@ -35,11 +36,17 @@ func CreateLeaseResponse(redboxLease *db.RedboxLease) *LeaseResponse {
 type LeaseResponse struct {
 	AccountID                string         `json:"accountId"`
 	PrincipalID              string         `json:"principalId"`
+	ID                       string         `json:"id"`
 	LeaseStatus              db.LeaseStatus `json:"leaseStatus"`
+	LeaseStatusReason        string         `json:"leaseStatusReturn"`
 	CreatedOn                int64          `json:"createdOn"`
 	LastModifiedOn           int64          `json:"lastModifiedOn"`
 	BudgetAmount             float64        `json:"budgetAmount"`
 	BudgetCurrency           string         `json:"budgetCurrency"`
 	BudgetNotificationEmails []string       `json:"budgetNotificationEmails"`
 	LeaseStatusModifiedOn    int64          `json:"leaseStatusModifiedOn"`
+	RequestedLeaseStart      int64          `json:"requestedLeaseStart"`
+	ActualLeaseStart         int64          `json:"actualLeaseStart"`
+	RequestedLeaseEnd        int64          `json:"requestedLeaseEnd"`
+	ActualLeaseEnd           int64          `json:"actualLeaseEnd"`
 }